avogadro/

+++ avogadro/io/trrformat.cpp
@@ -62,10 +62,10 @@ int isDouble(map<string, int>& header)
   for (auto& headerKey : headerKeys) {
     if (header[headerKey] != 0) {
       if (headerKey == "box_size") {
-        size = (int)(header[headerKey] / DIM * DIM);
+        size = header[headerKey] / DIM * DIM;
         break;
       } else {
-        size = (int)(header[headerKey] / (header["natoms"] * DIM));
+        size = header[headerKey] / (header["natoms"] * DIM);
         break;
       }
     }
